Rename filterBuilder and filterBuildingView to match casing conventions
