+++ src/views/filterbuilding/FilterBuildingView.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Header from "../../components/header/Header";
 import "../../components/alerttable/alerttable.css";
-import FilterBuilder from "../../components/filterbuilder/filterBuilder";
+import FilterBuilder from "../../components/filterbuilder/FilterBuilder";
 import { withRouter } from "react-router-dom";
 type PropType = {
   history: any;
